requirements.txt
requests>=2.25.1
pytest>=6.2.5
pytest-html>=3.1.1
jsonschema>=4.0.0
PyYAML>=6.0
Jinja2>=3.0.0
Appium-Python-Client>=2.7.1
selenium>=4.0.0
pytest-xdist>=2.5.0
allure-pytest>=2.9.45
